+++ integration_tests/juniper_tests/src/codegen/mod.rs
@@ -2,5 +2,6 @@ mod derive_enum;
 mod derive_input_object;
 mod derive_object;
 mod derive_object_with_raw_idents;
+mod derive_union;
 mod impl_union;
 mod scalar_value_transparent;